+++ rate.go
@@ -55,6 +55,27 @@ type SmartRate struct {
 	BillingType            string         `json:"billing_type,omitempty"`
 }
 
+// A StatelessRate contains information on shipping cost and delivery time, but does not have an ID (is ephemeral).
+type StatelessRate struct {
+	Object                 string     `json:"object,omitempty"`
+	Mode                   string     `json:"mode,omitempty"`
+	Service                string     `json:"service,omitempty"`
+	Carrier                string     `json:"carrier,omitempty"`
+	CarrierAccountID       string     `json:"carrier_account_id,omitempty"`
+	ShipmentID             string     `json:"shipment_id,omitempty"`
+	Rate                   string     `json:"rate,omitempty"`
+	Currency               string     `json:"currency,omitempty"`
+	RetailRate             string     `json:"retail_rate,omitempty"`
+	RetailCurrency         string     `json:"retail_currency,omitempty"`
+	ListRate               string     `json:"list_rate,omitempty"`
+	ListCurrency           string     `json:"list_currency,omitempty"`
+	DeliveryDays           int        `json:"delivery_days,omitempty"`
+	DeliveryDate           *time.Time `json:"delivery_date,omitempty"`
+	DeliveryDateGuaranteed bool       `json:"delivery_date_guaranteed,omitempty"`
+	EstDeliveryDays        int        `json:"est_delivery_days,omitempty"`
+	BillingType            string     `json:"billing_type,omitempty"`
+}
+
 // TimeInTransit provides details on the probability your package will arrive within a certain number of days
 type TimeInTransit struct {
 	Percentile50 int `json:"percentile_50,omitempty"`