+++ tests/rate_test.go
@@ -20,3 +20,27 @@ func (c *ClientTests) TestRateRetrieve() {
 	assert.Equal(reflect.TypeOf(&easypost.Rate{}), reflect.TypeOf(rate))
 	assert.True(strings.HasPrefix(rate.ID, "rate_"))
 }
+
+func (c *ClientTests) TestBetaStatelessRateRetrieve() {
+	client := c.TestClient()
+	assert, require := c.Assert(), c.Require()
+
+	rates, err := client.BetaGetStatelessRatesForShipment(c.fixture.BasicShipment())
+	require.NoError(err)
+
+	assert.Equal(reflect.TypeOf([]*easypost.StatelessRate{}), reflect.TypeOf(rates))
+}
+
+func (c *ClientTests) TestBetaStatelessRateGetLowest() {
+	client := c.TestClient()
+	assert, require := c.Assert(), c.Require()
+
+	rates, err := client.BetaGetStatelessRatesForShipment(c.fixture.BasicShipment())
+	require.NoError(err)
+
+	lowestRate, err := client.LowestStatelessRate(rates)
+	require.NoError(err)
+
+	assert.Equal(reflect.TypeOf(easypost.StatelessRate{}), reflect.TypeOf(lowestRate))
+	assert.Equal("First", lowestRate.Service)
+}